% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_derivative.R
\name{add_derivative}
\alias{add_derivative}
\title{Adds derivatives of C and Psi to the internal list}
\usage{
add_derivative(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{build_C} returns the inputted internal_list. The slot "derivative" of
the C and Psi slot is populated with a numerical matrix consisting of zeros and ones.
}
\description{
Internal function that adds the partial derivative of the vectorized
C and Psi matrices with respect to the parameters.
   verbose argument is written into the internal list
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\seealso{
\code{\link{build_Psi}}
}
\keyword{internal}
