% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels_in_lavaan_parTable.R
\name{add_labels_in_lavaan_parTable}
\alias{add_labels_in_lavaan_parTable}
\title{Fill in Labels in lavaan parTable Object}
\usage{
add_labels_in_lavaan_parTable(internal_list)
}
\arguments{
\item{internal_list}{a list with various information extracted 
from the model.}
}
\value{
A lavaan parTable object where empty slots (i.e., parameters 
that do not have a user specified label) in the original 
parTable object have been filled in with default labels from 
the lavaan model syntax.
}
\description{
Fills in default labels from the lavaan model 
syntax into empty slots (i.e., parameters that do not have a user
specified label) of the lavaan parTable object.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
