% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_C.R
\name{build_C}
\alias{build_C}
\title{Extracts the structural coefficient matrix from a fitted SEM object}
\usage{
build_C(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{build_C} returns the inputted internal_list with slot
   C populated with a two-entry list: "values" is a numeric matrix
   containing the structural coefficients (if available with row- and
   column names); "labels" is a character matrix containing parameter
   labels (NA for unlabeled parameters), NULL if labels are not extractable.
}
\description{
Internal function that extracts the structural coefficient matrix from a fitted
   structural equation model. Supported fitted objects: lavaan.
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\seealso{
\code{\link{build_Psi}}
}
\keyword{internal}
