% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_psi.R
\name{build_Psi}
\alias{build_Psi}
\title{Extracts the Psi matrix from a SEM}
\usage{
build_Psi(internal_list)
}
\arguments{
\item{internal_list}{a list with various information extracted from the
model.}
}
\value{
\code{build_psi} returns a list with various information extracted
from the model. \code{build_psi} with changed 'values' and 'labels' slots of Psi.
}
\description{
Internal function that extracts the Psi matrix from a lavaan
model. Currently, this function will only work if all variance parameters
are stored in lavaan's psi matrix (covariance of the latent errors) and the
theta matrix (covariance of the manifest errors) is empty.
}
\references{
Gische, C. & Voelkle, M. C. (under review???). Beyond the mean: A flexible framework for studying causal effects using linear models
\href{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
