% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_theta.R
\name{build_theta}
\alias{build_theta}
\title{Extracts the number, the numeric values, and the labels
of distinct and functionally unrelated SEM model parameters}
\usage{
build_theta(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{build_theta} returns the inputted internal_list with slot
   param populated with a five-entry list:
   "n_par" is an integer indicating the total number of estimated parameters
    (potential duplicates due to symmetry or equality constraints are counted).
   "n_par_unique" is an integer indicating the number of distinct and functionally
    unrelated parameters.
   "labels_par_unique" is a character vector containing the labels of distinct
    and functionally unrelated parameters.
   "values_par_unique" is a numeric vector containing the parameter values
    (estimates) of distinct and functionally unrelated parameters.
   "varcov_par_unique" is a numeric matrix containing equal to the variance-covariance
     matrix of the estimator of distinct and functionally unrelated parameters.
}
\description{
Internal function that extracts the numeric values and
  labels of distinct and functionally unrelated parameters from a
  fitted structural equation model (i.e., all uniquely labelled
  coefficients in C and Psi matrix. Duplicates due to symmetry or
  equality constraints are removed.). The entries in the resulting
  vectors follow the following predefined order:
  First: structural coefficients from the C matrix are added rowwise.
  Second: variance covariance parameters from the Psi matrix are added rowwise.
  Supported fitted objects: lavaan.
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\seealso{
\code{\link{build_Psi, build_C}}
}
\keyword{internal}
