% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_zero_one_matrix.R
\name{build_zero_one_matrix}
\alias{build_zero_one_matrix}
\title{Builds zero-one matrices (see Definition 1 in Gische and Voelkle, 2021) for the computation of the interventional distribution}
\usage{
build_zero_one_matrix(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{build_zero_one_matrix} returns the inputted internal_list with several slots
   in ..$zero_one_matrices filled. 
   ..$select_intervention: 
   ..$select_non_intervention:
   ..$eliminate_intervention:
}
\description{
Internal function that build zero-one matrices (e.g., selection matrices) for the computation of the interventional distribution
}
\references{
Gische, C. & Voelkle, M. C. (2021). Beyond the mean: A flexible framework for
   studying causal effects using linear models. Psychometrika.
}
\seealso{
\code{\link{intervention_moments}}
}
\keyword{internal}
