% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ase_probability.R
\name{calc_ase_probability}
\alias{calc_ase_probability}
\title{Calculate asymptotic standard errors of the interventional
distribution}
\usage{
calc_ase_probability(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{calc_ase_probability} returns the inputted internal.list and adds
three slots to the interventional_distribution$probability sublist.
   "acov" is a numeric with the asymptotic variance.
   "ase" is a numeric with the asymptotic standard error.
   "zvalue" is a numeric with the asymptotic z-value.
}
\description{
The function ase_probability() calculates the asysmptotic variance,
standard error and z-value of the interventional probability. Currently, the
function only accepts a single outcome variable.
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
