% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_interventional_probability.R
\name{calc_interventional_probability}
\alias{calc_interventional_probability}
\title{Function to calculate interventional probability}
\usage{
calc_interventional_probability(mean, sd, y_low, y_up, verbose)
}
\arguments{
\item{mean}{mean(s)}

\item{sd}{standard deviation(s)}

\item{y_low}{lower bound(s)}

\item{y_up}{upper bound(s)}

\item{verbose}{verbosity of console outputs}
}
\value{
\code{calc_interventional_probability} returns the
   interventional probability/ies (numeric value/s) as defined in Eq. 22d (p. 22)
}
\description{
Internal function that calculates interventional probability
   (can be used with scalars or vectors as inputs)
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\keyword{internal}
