% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ase.R
\name{calculate_ase}
\alias{calculate_ase}
\title{Calculate asymptotic standard errors}
\usage{
calculate_ase(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{calculate_ase} returns the inputted internal_list and adds
three slots to the interventional_distribution sublist.
   "acov" is a list containing the asymptotic variance-covariance matrices of
   the interventional means and covariances.
   "ase" is a list containing vectors with the asymptotic standard errors of
   the interventional means and variances.
   "z_values" is a list containing vectors with the asymptotic z-values of
   the interventional means and variances.
}
\description{
Internal function that calculates the asymptotic
variance-covariance matrix, standard errors, and z-values of the
interventional means and variance-covariance matrix
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
