% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ase_interventional_density.R
\name{calculate_ase_interventional_density}
\alias{calculate_ase_interventional_density}
\title{Calculate Asymptotics of the Interventional Probability Density 
Function}
\usage{
calculate_ase_interventional_density(
  model,
  x,
  y,
  intervention_names,
  outcome_names,
  verbose
)
}
\arguments{
\item{model}{An object of class causalSEM (e.g., the output of the 
intervention_effect() function)}

\item{x}{interventional levels}

\item{y}{values within the range of the outcome variable}

\item{intervention_names}{names of interventional variables}

\item{outcome_names}{names of outcome variables}

\item{verbose}{verbosity of console outputs}
}
\value{
A list containing the asysmptotic covariance matrix, 
the aysmptotic standard errors, and the approximate z-value of the 
the interventional pdf for a specific interventional level and a specific 
value in the range of the outcome variable.
(see Corollaries, 10, 11 in Gische and Voelkle, 2021)
}
\description{
Calculates the asysmptotic variance,
the aysmptotic standard error, and the approximate z-value of the  
interventional probability density function for a specific interventional 
level and a specific value in the range of a univariate outcome variable.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
