% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ase_interventional_means.R
\name{calculate_ase_interventional_means}
\alias{calculate_ase_interventional_means}
\title{Calculate Asymptotics of the Interventional Mean Vector}
\usage{
calculate_ase_interventional_means(
  model,
  x,
  intervention_names,
  outcome_names,
  verbose
)
}
\arguments{
\item{model}{internal_list or object of class causalSEM}

\item{x}{interventional levels}

\item{intervention_names}{names of interventional variables}

\item{outcome_names}{names of outcome variables}

\item{verbose}{verbosity of console outputs}
}
\value{
The asysmptotic covariance matrix, the aysmptotic 
standard errors, and the approximate z-values of the the interventional
means for a specific interventional level.
(see Corollaries, 10, 11 in Gische and Voelkle, 2021).
}
\description{
Calculates the asysmptotic covariance matrix,
the aysmptotic standard errors, and the approximate z-values 
of the of the interventional means for a specific interventional level.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
