% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ase_interventional_probabilities.R
\name{calculate_ase_interventional_probabilities}
\alias{calculate_ase_interventional_probabilities}
\title{Calculate Asymptotic Standard Errors of Interventional Probabilities}
\usage{
calculate_ase_interventional_probabilities(
  model,
  x,
  intervention_names,
  outcome_names,
  lower_bounds,
  upper_bounds,
  verbose
)
}
\arguments{
\item{model}{internal_list or object of class causalSEM}

\item{x}{interventional levels}

\item{intervention_names}{names of interventional variables}

\item{outcome_names}{names of outcome variables}

\item{lower_bounds}{numeric with lower bounds}

\item{upper_bounds}{numeric with upper bounds}

\item{verbose}{verbosity of console outputs}
}
\value{
Asysmptotic covariance matrix, the aysmptotic standard errors, and
the approximate z-values of the the interventional interventional  
probabilities for a specific interventional level and a specific range 
(lower_bounds, upper_bounds) of the outcome variables. 
(see Corollaries, 10, 11 in Gische and Voelkle, 2021).
}
\description{
Calculates the asysmptotic covariance matrix,
the aysmptotic standard errors, and the approximate z-values of the 
interventional probabilities for a specific interventional level and a 
specific range (lower_bounds, upper_bounds) of the outcome variables.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
