% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ase_interventional_variances.R
\name{calculate_ase_interventional_variances}
\alias{calculate_ase_interventional_variances}
\title{Calculate Asymptotics of the Interventional Variances}
\usage{
calculate_ase_interventional_variances(
  model,
  x,
  intervention_names,
  outcome_names,
  verbose
)
}
\arguments{
\item{model}{internal_list or object of class causalSEM}

\item{x}{interventional level}

\item{intervention_names}{names of interventional variables}

\item{outcome_names}{name of outcome variable}

\item{verbose}{verbosity of console outputs}
}
\value{
The asysmptotic covariance matrix, the aysmptotic standard errors, 
and the approximate z-values of the the interventional variances for a 
specific interventional level.
(see Corollaries, 10, 11 in Gische and Voelkle, 2021).
}
\description{
Calculates the asysmptotic covariance matrix,
the aysmptotic standard errors, and the approximate z-values 
of the of the interventional variances for a specific interventional level.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
