% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interventional_probabilities.R
\name{calculate_interventional_probabilities}
\alias{calculate_interventional_probabilities}
\title{Calculate Probabilites of Interventional Events}
\usage{
calculate_interventional_probabilities(mean, sd, y_low, y_up, verbose)
}
\arguments{
\item{mean}{mean(s)}

\item{sd}{standard deviation(s)}

\item{y_low}{lower bound(s)}

\item{y_up}{upper bound(s)}

\item{verbose}{verbosity of console outputs}
}
\value{
The interventional probability/ies (numeric value/s) as defined in
 Eq. 22d (p. 22) of Gische and Voelkle (2021).
}
\description{
Calculates probabilities of interventional events, for example,
the probabily that the outcome variable realizes within a critical range of
values given an intervention on the exposure.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\keyword{internal}
