% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jacobian_C.R
\name{calculate_jacobian_C}
\alias{calculate_jacobian_C}
\title{Calculate Jacobian of C-Matrix}
\usage{
calculate_jacobian_C(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted list with the slot \code{..$derivative}
 in the sublist \code{..$C} filled in:
\tabular{lll}{
     \code{..$derivative}:\tab \tab Numeric matrix containing 
     the Jacobian of the vectorized C matrix.\cr}
}
\description{
Calculates the Jacobian of the vectorized matrix of structural 
coefficients (C-matrix). The Jacobian is defined as the partial derivative 
with respect to the parameters vector theta of distinct and functionally 
unrelated parameters.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
