% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jacobian_interventional_density.R
\name{calculate_jacobian_interventional_density}
\alias{calculate_jacobian_interventional_density}
\title{Calculate the Jacobian of the Density Function of the Interventional 
Distribution}
\usage{
calculate_jacobian_interventional_density(
  model,
  x,
  y,
  intervention_names,
  outcome_names,
  verbose
)
}
\arguments{
\item{model}{internal_list or object of class causalSEM}

\item{x}{interventional level}

\item{y}{value in the range of the outcome variable}

\item{intervention_names}{names of interventional variables}

\item{outcome_names}{name of outcome variable}

\item{verbose}{verbosity of console outputs}
}
\value{
\code{calculate_jacobian_interventional_density} returns the
   Jacobian of interventional probability density function (numeric value) 
   as defined in Eq. 18c (p. 17)
}
\description{
Calculate the Jacobian of the probability density function (pdf)
of the Interventional distributionfor a specific interventional level and a 
specific value in the range of the outcome variable.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\keyword{internal}
