% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jacobian_interventional_means.R
\name{calculate_jacobian_interventional_means}
\alias{calculate_jacobian_interventional_means}
\title{Calculate Jacobian of the Interventional Mean Vector}
\usage{
calculate_jacobian_interventional_means(
  model,
  x,
  intervention_names,
  outcome_names,
  verbose
)
}
\arguments{
\item{model}{internal_list or object of class causalSEM}

\item{x}{interventional level}

\item{intervention_names}{names of interventional variables}

\item{outcome_names}{name of outcome variable}

\item{verbose}{verbosity of console outputs}
}
\value{
The Jacobian of interventional mean vector (numeric matrices)
   as defined in Eq. 18a (p. 17).
}
\description{
Calculate Jacobian of the interventional mean Vector for a 
specific interventional level.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\keyword{internal}
