% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jacobian_interventional_variances.R
\name{calculate_jacobian_interventional_variances}
\alias{calculate_jacobian_interventional_variances}
\title{Calculate Jacobian of the Covariance Matrix of the Interventional
Distribution}
\usage{
calculate_jacobian_interventional_variances(
  model,
  intervention_names,
  outcome_names,
  verbose
)
}
\arguments{
\item{model}{internal_list or object of class causalSEM}

\item{intervention_names}{names of interventional variables}

\item{outcome_names}{names of outcome variable}

\item{verbose}{verbosity of console outputs}
}
\value{
The Jacobian of interventional covariance matrix (numeric matrices)
   as defined in Eq. 18b (p. 17)
}
\description{
Calculate Jacobian of the covariance matrix of the
interventional distribution for a specific interventional level.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
