% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jacobian_Psi.R
\name{calculate_jacobian_Psi}
\alias{calculate_jacobian_Psi}
\title{Calculate Jacobian of Covariance Matrix of Error Terms}
\usage{
calculate_jacobian_Psi(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted internal_list with the slot "derivative" of
the ..$Psi slot containing the vectorized Jacobian of the Psi matrix.
}
\description{
Calculates the Jacobian of the vectorized covariance matrix of  
error terms (partial derivative of Psi with respect to the parameters vector
theta).
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\seealso{
\code{\link{calculate_jacobian_C}}
}
\keyword{internal}
