% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.causalSEM.R
\name{coef.causalSEM}
\alias{coef.causalSEM}
\title{Extract Causal Coefficients}
\usage{
\method{coef}{causalSEM}(x, what = "means")
}
\arguments{
\item{x}{an object of class "causalSEM", usually, a result of a call to intervention_effect.}

\item{what}{string indicating what type of coefficients are to be returned.
'means' returns interventional means, 'var' returns interventional variances,
and 'varcov' returns interventional variances and covariances.}
}
\value{
\code{coef.causalSEM} numeric vector.
}
\description{
coef method for class "causalSEM".
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
