% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_causalSEM_s3_object.R
\name{create_causalSEM_s3_object}
\alias{create_causalSEM_s3_object}
\title{causalSEM S3 object}
\usage{
create_causalSEM_s3_object(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
object of class causalSEM
}
\description{
Assign class causalSEM to internal list
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\keyword{internal}
