% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_list.R
\name{create_empty_list}
\alias{create_empty_list}
\title{Create Empty Internal List}
\usage{
create_empty_list(verbose = NULL)
}
\arguments{
\item{verbose}{Integer number describing verbosity of console output.
0...no output (default); 1...user messages; 2...debugging-relevant messages.}
}
\value{
An empty internal list with the following structure:\cr
 \tabular{lll}{
    List of 6\cr
    \code{$fitted_object}: List       \tab \tab Object containing fitted 
    model. \cr
    \code{$fitted_object_class}: chr(0) \tab \tab Class of fitted model. \cr
    \code{$info_model}: List of 6  \tab \tab \cr
     \code{..$n_obs}: int(0)       \tab \tab Number of observations.\cr
     \code{..$n_ov}: int(0)        \tab \tab Number of manifest variables.\cr
     \code{..$var_names}: chr(0)   \tab \tab Names of observed variables.\cr
     \code{..$C}: List of 3        \tab \tab \cr
     \code{.. ..$values}: num[0, 0]   \tab \tab Parameter values of matrix of 
     structural coefficients.\cr
     \code{.. ..$labels}: chr[0, 0]   \tab \tab Parameter labels of matrix of 
     structural coefficients.\cr
     \code{.. ..$derivative}: chr[0, 0]  \tab \tab Partial derivative of the 
     vectorized C matrix with \cr
                                \tab \tab respect to the parameters.\cr
     \code{..$ Psi}: List of 3         \tab \tab \cr
     \code{.. ..$values}: num[0, 0]   \tab \tab Parameter values of model 
     implied covariance matrix.\cr
     \code{.. ..$labels}: chr[0, 0]   \tab \tab Parameter labels of model 
     implied covariance matrix\cr
     \code{.. ..$derivative}: chr[0, 0]  \tab \tab Partial derivative of the 
     vectorized Psi matrix with\cr
                                \tab \tab respect to the parameters.\cr
     \code{..$param}: List of 5       \tab \tab \cr
     \code{.. ..$n_par}: int(0)        \tab \tab Total number of estimated 
     parameters.\cr
     \code{.. ..$n_par_unique}: int(0) \tab \tab Number of distinct and 
     functionally unrelated parameters.\cr
     \code{.. ..$labels_par_unique}: chr(0)  \tab \tab Labels of distinct and 
     functionally unrelated parameters.\cr
     \code{.. ..$values_par_unique}: num(0)  \tab \tab Values of distinct
      and functionally unrelated parameters.\cr
     \code{.. ..$varcov_par_unique}: num[0, 0] \tab \tab Covariance 
     matrix of the estimator of distinct\cr
                         \tab \tab and functionally unrelated parameters.\cr
    \code{$info_interventions}: List of 8      \tab \tab \cr
     \code{..$n_intervention}: int(0)    \tab \tab Number of interventional 
     variables.\cr
     \code{..$intervention_names}: chr(0) \tab \tab Names of interventional 
     variables.\cr
   \code{..$n_outcome}: int(0)   \tab \tab Number of outcome variables.\cr
   \code{..$outcome_names}: chr(0) \tab \tab Names of outcome variables.\cr
   \code{..$intervention_levels}: num(0) \tab \tab Interventional levels.\cr
   \code{..$effect_type}: chr(0)   \tab \tab Features of the interventional 
     distribution to be analyzed.\cr
     \code{..$lower_bound}: num(0)   \tab \tab Lower bound of critical range 
     of outcome variables.\cr
     \code{..$upper_bound}: num(0)   \tab \tab Upper bound of critical range
      of outcome variables.\cr
    \code{$matrices_constant}: List of 7   \tab \tab \cr
     \code{.. .. $select_intervention}: num[0, 0]  \tab \tab Selection matrix 
     for entries that are intervened on. \cr
     \code{.. .. $select_non_intervention}: num[0, 0] \tab \tab Selection  
     matrix for entries that are NOT intervened on. \cr
     \code{.. .. $select_outcome}: num[0, 0]          \tab \tab Selection 
     matrix for outcomes of interest. \cr
     \code{.. .. $eliminate_intervention}: num[0, 0]  \tab \tab Matrix that 
     replaces entries that are intervened on by zero. \cr
     \code{.. .. $duplication_matrix}: num[0, 0]      \tab \tab Maps vech(A)  
     onto vec(A) for symmetric A. \cr
     \code{.. .. $elimination_matrix}: num[0, 0]      \tab \tab Maps vec(A) 
     onto vech(A). \cr
     \code{.. .. $commutation_matrix}: num[0, 0]      \tab \tab Maps vech(A) 
     onto vec(A'). \cr
    \code{$interventional_distribution} : List of 4   \tab \tab \cr
     \code{..$means}: List of 4        \tab \tab \cr
     \code{.. .. $values}: num(0)      \tab \tab Mean vector of the  
     interventional distribution. \cr
     \code{.. .. $jacobian}: num[0, 0] \tab \tab Jacobian of interventinoal 
     mean vector. \cr
     \code{.. .. $ase}: num(0)       \tab \tab Asymptotic standard errors of 
     interventional means.  \cr
     \code{.. .. $z_values}: num[0, 0] \tab \tab z-values of interventional 
     means.\cr
     \code{..$covariance_matrix}: List of 4    \tab \tab \cr
     \code{.. .. $values}: num(0)      \tab \tab Covariance matrix of the 
     interventional distribution.  \cr
     \code{.. .. $jacobian}: num[0, 0] \tab \tab Jacobian of VECTORIZED 
     interventional covariance matrix. \cr
     \code{.. .. $ase}: num(0)        \tab \tab Asymptotic standard errors of 
     interventional (co)variances.  \cr
     \code{.. .. $z_values}: num[0, 0] \tab \tab z-values of interventional 
     interventional (co)variances. \cr
     \code{..$density_function}: List of 4    \tab \tab \cr
     \code{.. .. $values}: num(0)      \tab \tab Probability density function 
     (pdf) of the\cr
     \tab \tab interventional distribution.  \cr
     \code{.. .. $jacobian}: num[0, 0] \tab \tab Jacobian of the 
     interventional pdf.\cr
     \code{.. .. $ase}: num(0)       \tab \tab asymptotic standard errors of 
     interventional pdf.   \cr
     \code{.. .. $z_values}: num[0, 0] \tab \tab z-values of interventional 
     pdf. \cr
     \code{..$probabilities}: List of 4  \tab \tab \cr
     \code{.. .. $values}: num(0)      \tab \tab Probabilities of  
     interventional events. \cr
     \code{.. .. $jacobian}: num[0, 0] \tab \tab Jacobian of interventional 
     probabilities. \cr
     \code{.. .. $ase}: num(0)         \tab \tab Asymptotic standard error of 
     interventional probabilities.  \cr
     \code{.. .. $z_values}: num[0, 0] \tab \tab z-value of the 
     interventional probabilities. \cr
    \code{$control}: List of 1     \tab \tab \cr
     \code{..$verbose}: num(0)     \tab \tab Verbosity of console output.\cr
    \code{$tables}: List of 2      \tab \tab \cr
     \code{..$interventional_means}: data.frame  \tab \tab Information for 
     \code{\link{summary.causalSEM}} \cr 
     \tab \tab and \code{\link{print.causalSEM}}  functions.\cr
     \code{..$interventional_means}: data.frame  \tab \tab Information for 
     \code{\link{summary.causalSEM}}\cr 
     \tab \tab and \code{\link{print.causalSEM}} 
     functions.\cr
 }
}
\description{
Creates an empty internal list of predefined structure. The 
list will subsequently be filled while the function  
\code{\link{intervention_effect}} is running.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
