% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{d_100}
\alias{d_100}
\title{Example data}
\format{
A data frame with 100 rows and 6 variables
}
\usage{
data(d_100)
}
\description{
A dataset containing some data. The variables are as follows:
}
\details{
\itemize{
  \item x1
  \item y1
  \item x2
  \item y2
  \item x3
  \item y3
}
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\keyword{datasets}
