% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_asymptotics_interventional_density.R
\name{fill_in_asymptotics_interventional_density}
\alias{fill_in_asymptotics_interventional_density}
\title{Fill in Asymptotics for Density Function of Interventional 
Distributions}
\usage{
fill_in_asymptotics_interventional_density(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted internal_list with slot 
interventional_distribution$density_function
..$ase
..$z-values
filled in with a list (one named entry for each variable). 
Each list element in ..$ase contains a matrix with two columns and 200 rows:
   [1] "x": generated x values (-3*SD ... 3*SD) 
   [2] "ase": asymptotic standard errors
Each list element in ..$z_values contains a matrix with two columns and 200 
rows:
   [1] "x": generated x values (-3*SD ... 3*SD) 
   [2] "z_values": asymptotic standard errors
}
\description{
Fill in asymptotics for density function of interventional 
distributions. Asymptotic values are computed pointwise over a grid of 200 
values in the range [-3*SD, 3*SD] of the univariate outcome variable. The 
asysmptotic standard errors and the approximate z-values of the  
interventional density function for a specific interventional level are 
computed pointwise over the grid.
}
\references{


Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
