% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_asymptotics_interventional_means.R
\name{fill_in_asymptotics_interventional_means}
\alias{fill_in_asymptotics_interventional_means}
\title{Fill in Asymptotics for the Mean Vector of the Interventional
 Distribution}
\usage{
fill_in_asymptotics_interventional_means(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the 
model.}
}
\value{
The inputted internal_list with several slots 
in ..$interventional_distribution filled in.
   ..$jacobian
   ..$acov
   ..$ase
   ..$z_value
}
\description{
Fills in the Jacobian, the asymptotic
covariance matrix, the asymptotic standard errors, and asymptotic z-values 
of the mean vector of the interventional distribution into the internal list.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\seealso{
\code{\link{}}
}
\keyword{internal}
