% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fill_in_asymptotics_interventional_probabilities.R
\name{fill_in_asymptotics_interventional_probabilities}
\alias{fill_in_asymptotics_interventional_probabilities}
\title{Fill in Asymptotic Quantities of Interventional Probabilities to 
Internal List}
\usage{
fill_in_asymptotics_interventional_probabilities(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the 
model.}
}
\value{
The inputted internal_list with several slots in 
..$interventional_distribution$probabilities filled.
   ..$jacobian
   ..$ase
   ..$z_values
}
\description{
Fills in the Jacobian, the asymptotic covariance matrix, 
the asymptotic standard errors, and asymptotic z-values of the
interventional probabilities into the internal list.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\seealso{
\code{\link{}}
}
\keyword{..$acov}
\keyword{..$ase}
\keyword{..$jacobian}
\keyword{..$z_values}
\keyword{internal}
