% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_constant_matrices.R
\name{fill_in_constant_matrices}
\alias{fill_in_constant_matrices}
\title{Fill in Zero-One Matrices to List}
\usage{
fill_in_constant_matrices(internal_list = NULL)
}
\arguments{
\item{internal_list}{A list with information extracted from the model.}
}
\value{
The inputted list with several slots in \code{..$constant_matrices} 
filled in.\cr
\tabular{ll}{
\tab   \code{$select_intervention} \cr
\tab   \code{$select_non_intervention} \cr
\tab   \code{$select_outcome} \cr
\tab   \code{$eliminate_intervention} \cr
\tab   \code{$duplication_matrix} \cr
\tab   \code{$elimination_matrix} \cr
\tab   \code{$commutation_matrix}}
}
\description{
Fill in constant zero-one matrices used for the computation of 
the interventional distribution into the internal list (see, for example, 
Definition 1 in Gische and Voelkle, 2022).
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
