% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_info_interventions.R
\name{fill_in_info_interventions}
\alias{fill_in_info_interventions}
\title{Process Arguments of the \code{intervention_effect}() Function}
\usage{
fill_in_info_interventions(
  internal_list = NULL,
  intervention = NULL,
  outcome = NULL,
  intervention_level = NULL,
  effect_type = NULL,
  lower_bound = NULL,
  upper_bound = NULL
)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted list with the following slots filled in:
\tabular{ll}{
List of 7 \tab \cr
\tab   \code{n_intervention} \cr
\tab   \code{intervention_names} \cr
\tab   \code{intervention_levels} \cr
\tab   \code{effect_type} \cr
\tab   \code{n_outcome} \cr
\tab   \code{outcome_names} \cr
\tab   \code{lower_bounds}\cr
\tab   \code{upper_bounds}}
}
\description{
Extract information about the intervention to be analyzed 
from the arguments of the \code{intervention_effect}() function.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
