% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_info_model.R
\name{fill_in_info_model}
\alias{fill_in_info_model}
\title{Extract Information From a Fitted Structural Equation Model Object}
\usage{
fill_in_info_model(internal_list, model)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}

\item{model}{Fitted model. The fitted model can be of class lavaan.}
}
\value{
The inputted internal_list with slots
   \code{n_obs}, \code{n_ov}, and \code{var_names} filled in.
}
\description{
Extract number of observations, number of manifest
   variables, and names of observed variables from a fitted
   structural equation model. Supported fitted objects: lavaan.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\seealso{
\code{\link{build_C}} \code{\link{build_Psi}}
}
\keyword{internal}
