% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_interventional_means.R
\name{fill_in_interventional_means}
\alias{fill_in_interventional_means}
\title{Fills in Mean Vector of the Interventional Distribution to Internal 
List}
\usage{
fill_in_interventional_means(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted internal_list with slot
   interventional_distribution$means$values filled in.
}
\description{
Fills in the mean vector of the interventional distribution to
the slot internal_list$interventional_distribution$means$values 
in the internal_list.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\keyword{internal}
