% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_interventional_probabilities.R
\name{fill_in_interventional_probabilities}
\alias{fill_in_interventional_probabilities}
\title{Fill in Interventional Probabilities to List}
\usage{
fill_in_interventional_probabilities(internal_list = NULL)
}
\arguments{
\item{internal_list}{A list with information extracted from the model.}
}
\value{
The inputted list with slot
\code{..$interventional_distribution$probabilities$values} 
filled in.
}
\description{
Fills in probabilities of interventional events to internal 
list. See, for example, Eqs. 10, 15, and 22d in Gische and Voelkle (2022).
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
