% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_jacobian_interventional_means.R
\name{fill_in_jacobian_interventional_means}
\alias{fill_in_jacobian_interventional_means}
\title{Fill in Jacobian for the Mean to List}
\usage{
fill_in_jacobian_interventional_means(internal_list = NULL)
}
\arguments{
\item{internal_list}{A list with various information extracted from the 
model.}
}
\value{
The inputted list with slots 
in \code{..$interventional_distribution$means} filled in.\cr
\tabular{lll}{
\tab   \code{..$jacobian} \tab The Jacobian matrix.}
}
\description{
Fills in the Jacobian of the mean vector of the interventional 
distribution into the internal list. See, for example, Theorem 9 and 
Corollaries 10 and 11 in Gische and Voelkle (2022).
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
