% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_print_table.R
\name{fill_in_print_table}
\alias{fill_in_print_table}
\title{Create output table with interventional means and variances}
\usage{
fill_in_print_table(internal_list)
}
\arguments{
\item{internal_list}{internal_list or object of class causalSEM}
}
\value{
\code{fill_in_print_table} returns a list containing two tables with
the interventional means and variances.
}
\description{
Internal function that creates a table with the interventional
means and variances with corresponding standard errors and 95% confidence
intervals.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
