% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_theta.R
\name{fill_in_theta}
\alias{fill_in_theta}
\title{Extract Model Parameters}
\usage{
fill_in_theta(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted internal_list with slots in the sublist \code{..$param}
filled in:
\tabular{lll}{
.. ..$n_par: \tab \tab Integer number indicating the total number of 
estimated parameters.\cr
\tab \tab Duplicates due to symmetry or equality constraints are counted.\cr
.. ..$n_par_unique  \tab \tab Integer number indicating the number of 
distinct and functionally unrelated parameters.\cr
.. ..$labels_par_unique  \tab \tab Character vector containing the labels of
 distinct and functionally unrelated parameters.\cr
.. ..$values_par_unique  \tab \tab Numeric vector containing the parameter 
values (estimates) of distinct and functionally\cr
\tab \tab unrelated parameters.\cr
.. ..$varcov_par_unique  \tab \tab Numeric matrix containing the 
(co-) variances of the estimator of distinct and\cr
 functionally unrelated parameters.}
}
\description{
Extracts the numeric values and
  labels of distinct and functionally unrelated parameters from a
  fitted structural equation model (i.e., all uniquely labelled
  coefficients in C and Psi matrix. Duplicates due to symmetry or
  equality constraints are removed.). The entries in the resulting
  vectors follow the following predefined order:
  First: structural coefficients from the C matrix are added rowwise.
  Second: variance covariance parameters from the Psi matrix are added 
  rowwise.
  Supported class of fitted objects: \code{lavaan}.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
