% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_effect.R
\name{intervention_effect}
\alias{intervention_effect}
\title{Calculate estimates of causal effects defined based on the interventional distribution.}
\usage{
intervention_effect(
  model,
  intervention,
  outcome = NULL,
  intervention_level,
  effect.type = NULL,
  lower.bound = NULL,
  upper.bound = NULL,
  verbose = 0,
  ...
)
}
\arguments{
\item{model}{Fitted model. The fitted model can be of class lavaan.}

\item{intervention}{Character vector. Variable names of interventional variables. Default: empty vector (no intervention).}

\item{outcome}{Character vector. Variable names of outcome variables. Default: all non-interventional variables.}

\item{intervention_level}{Numeric vector. Interventional levels. Same length and order as argument intervention. Default: vector of ones.}

\item{effect.type}{Character vector. Parts of the interventional distribution the user is interested in. Admissible values are "mean", "variance", "density", and "probability". Default: "mean". Argument "probability" only admissible if outcome is univariate.}

\item{lower.bound}{Single number, numeric. Lower bound of critical range of univariate outcome variable.}

\item{upper.bound}{Single number, numeric. Upper bound of critical range of univariate outcome variable.}

\item{verbose}{A single number, integer. 0...no output (default), 1...user messages, 2...debugging-relevant messages.}
}
\value{
An object of class causalSEM, for which several methods are available, including a summary method.
}
\description{
Calculate estimates of causal effects defined based on the interventional distribution.
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\keyword{extenal}
