% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventional_density.R
\name{interventional_density}
\alias{interventional_density}
\title{Calculates interventional density}
\usage{
interventional_density(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{interventional_density} returns the inputted internal_list with slot
   interventional_distribution$density$pdf populated with a list (one named entry for each variable),
   each list element contains a matrix with two columns and 200 rows:
   [1] "x": generated x values (-3*SD ... 3*SD) 
   [2] "pdf.values": values from dnorm
}
\description{
Internal function that calculates interventional density
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\keyword{internal}
