% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventional_moments.R
\name{interventional_moments}
\alias{interventional_moments}
\title{Calculates interventional moments}
\usage{
interventional_moments(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{interventional_moments} returns the inputted internal_list with slot
   interventional_distribution$moments$mean_vector and ..$variance_matrix populated
}
\description{
Internal function that calculates interventional moments
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\keyword{internal}
