% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventional_probability.R
\name{interventional_probability}
\alias{interventional_probability}
\title{Calculates interventional probability}
\usage{
interventional_probability(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{interventional_probability} returns the inputted internal_list with slot
   interventional_distribution$probability$p populated with (a) numeric value(s);
   interventional probabilities are only calculated for outcome variables 
   (drawn from internal_list$info_interventions$outcome_name)
}
\description{
Internal function that computes interventional probabilities for outcome variables
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\keyword{internal}
