% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_parTable_fill_labels.R
\name{lav_parTable_fill_labels}
\alias{lav_parTable_fill_labels}
\title{Fill in labels in lavaan parTable object}
\usage{
lav_parTable_fill_labels(internal_list)
}
\arguments{
\item{internal_list}{a list with various information extracted from the
model.}
}
\value{
\code{lav_parTable_fill_labels} returns a lavaan parTable object where empty slots (i.e., parameters that do not have a user specified label) in the original parTable object have been filled in with default labels from the lavaan model syntax.
}
\description{
Internal function that fills in default labels from the lavaan model syntax into empty slots (i.e., parameters that do not have a user specified label) of the lavaan parTable object.
}
\references{
Gische, C. & Voelkle, M. C. (under review???). Beyond the mean: A flexible framework for studying causal effects using linear models
\href{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
