% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_empty_list.R
\name{make_empty_list}
\alias{make_empty_list}
\title{Builds an empty internal list}
\usage{
make_empty_list(verbose = NULL)
}
\arguments{
\item{verbose}{A number, 0...no output (default), 1...user messages,
2...debugging-relevant messages. Valid default is
set by function \code{\link{verbose_argument_handling}}}
}
\value{
\code{make_empty_list} returns an empty internal list with structure:\cr
 \code{\tabular{lll}{
    List of 5\cr
    $ fitted_object      : NULL           \tab \tab \cr
    $ fitted_object_class: NULL           \tab \tab \cr
    $ info_model         :List of 10      \tab \tab \cr
     ..$ n_obs            : int 0         \tab \tab number of observations\cr
     ..$ n_ov             : int 0         \tab \tab number of manifest variables\cr
     ..$ var_names        : chr(0)        \tab \tab names of observed variables\cr
     ..$ C                :List of 2      \tab \tab \cr
     .. ..$ values: num[0 , 0 ]           \tab \tab parameter values of matrix of structural coefficients\cr
     .. ..$ labels: chr[0 , 0 ]           \tab \tab parameter names of matrix of structural coefficients\cr
     ..$ Psi              :List of 2      \tab \tab \cr
     .. ..$ values: num[0 , 0 ]           \tab \tab parameter values of model implied covariance matrix\cr
     .. ..$ labels: chr[0 , 0 ]           \tab \tab parameter names of model implied covariance matrix\cr
     ..$ n_par            : int 0         \tab \tab total number of estimated parameters\cr
     ..$ n_par_unique     : int 0         \tab \tab number of distinct and functionally unrelated parameters\cr
     ..$ par_names_unique : chr(0)        \tab \tab names of distinct and functionally unrelated parameters\cr
     ..$ par_unique       : num(0)        \tab \tab parameter values (estimates) of distinct and functionally unrelated parameters\cr
     ..$ varcov_par_unique: num[0 , 0 ]   \tab \tab variance-covariance matrix of estimator of vector\cr
    $ info_interventions :List of 8       \tab \tab \cr
     ..$ n_intervention    : int 0        \tab \tab number of interventional variables\cr
     ..$ intervention_name : chr(0)       \tab \tab names of interventional variables\cr
     ..$ n_outcome         : int 0        \tab \tab number of outcome variables\cr
     ..$ outcome_name      : chr(0)       \tab \tab names of outcome variables\cr
     ..$ intervention_level: num(0)       \tab \tab interventional level\cr
     ..$ effect_type       : chr(0)       \tab \tab parts of the interventional distribution the user is interested in\cr
     ..$ lower_bound       : num(0)       \tab \tab lower bound of critical range of univariate outcome variable\cr
     ..$ upper_bound       : num(0)       \tab \tab upper bound of critical range of univariate outcome variable\cr
    $ control            :List of 1       \tab \tab \cr
     ..$ verbose: num 0                   \tab \tab verbosity of console output\cr
 }}
}
\description{
Internal function that builds an empty internal list;
   verbose argument is written into the internal list
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\seealso{
\code{\link{verbose_argument_handling}}
}
\keyword{internal}
