% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot density}
\usage{
plot_density(object, plot = TRUE)
}
\arguments{
\item{object}{The object returned from calling function \code{intervention_effect}.}

\item{plot}{logical, if TRUE plots are displayed}
}
\value{
\code{plot_density} returns ggplot2 code for density plots
}
\description{
Function that generates density plots
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
