% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot distributions}
\usage{
plot_distributions(object, plot = TRUE, plot.dir = NULL)
}
\arguments{
\item{object}{The object returned from calling
function \code{intervention_effect}.}

\item{plot}{directory, directory to save pdf plot files}
}
\value{
\code{plot_distributions} returns ggplot2 code
        for density plots
}
\description{
Function that generates distribution plots
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a
            flexible framework for studying causal effects using linear
            models. Psychometrika (advanced online publication).
            https://doi.org/10.1007/s11336-021-09811-z
}
