% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_intervention_info.R
\name{populate_intervention_info}
\alias{populate_intervention_info}
\title{Extracts interventional information from the arguments of the intervention_effect function.}
\usage{
populate_intervention_info(
  internal_list,
  intervention,
  outcome,
  intervention_level,
  effect.type,
  lower.bound,
  upper.bound
)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{populate_intervention_info} returns the inputted internal_list with slots
   \code{n_intervention}, \code{intervention_name}, \code{intervention_level},
   \code{effect_type}, \code{n_outcome}, \code{outcome_name},  
   \code{lower_bound}, and \code{upper_bound} populated.
}
\description{
Internal function that Extracts interventional information from the arguments 
   of the intervention_effect function.
}
\references{
Gische, C. & Voelkle, M. C. (2022). Beyond the mean: A flexible framework for
   studying causal effects using linear models. Psychometrika.
}
\seealso{
\code{\link{populate_model_info}}
}
\keyword{internal}
