% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_model_info.R
\name{populate_model_info}
\alias{populate_model_info}
\title{Extracts number of observations, number of manifest variables,
   and from a names of observed variables from a fitted SEM object}
\usage{
populate_model_info(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
\code{populate_model_info} returns the inputted internal_list with slots
   \code{n_obs}, \code{n_ov}, and \code{var_names} populated.
}
\description{
Internal function that extracts number of observations,
   number of manifest variables, and names of observed variables from a fitted
   structural equation model. Supported fitted objects: lavaan.
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\seealso{
\code{\link{build_C}} \code{\link{build_Psi}}
}
\keyword{internal}
