% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_table_variances.R
\name{prepare_table_variances}
\alias{prepare_table_variances}
\title{Prepare Table for Interventional Variances}
\usage{
prepare_table_variances(x, digits = 3)
}
\arguments{
\item{x}{An object of class \code{causalSEM}.}

\item{digit}{Single number, integer. Indicating the number of decimal places
(round) or significant digits (signif) to be used.}
}
\value{
A character vector with information about interventional variances.
}
\description{
Create a formatted table containing information 
about interventional variances. Used in \code{print.causalSEM} and 
\code{summary.causalSEM}.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
