% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.causalSEM.R
\name{summary.causalSEM}
\alias{summary.causalSEM}
\title{Summary of Object of Class \code{causalSEM}}
\usage{
\method{summary}{causalSEM}(x, digits = 3)
}
\arguments{
\item{x}{An object of class \code{causalSEM}.}

\item{digit}{Single number, integer. Indicating the number of decimal places
(round) or significant digits (signif) to be used.}
}
\value{
A list summarizing selected information and statistics from an object 
of class \code{causalSEM}.
}
\description{
Summary method for objects of class \code{causalSEM}.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
