% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose_argument_handling.R
\name{verbose_argument_handling}
\alias{verbose_argument_handling}
\title{Checks and potentially "repairs" the verbose argument}
\usage{
verbose_argument_handling(verbose = c(0, 1, 2))
}
\arguments{
\item{verbose}{A number, 0...no output (default), 1...user messages,
2...debugging-relevant messages.}
}
\value{
returns either the valid user input or 0 (no output) as default
}
\description{
Internal function, checks whether the user-inputted verbose
   argument is valid
}
\keyword{internal}
