% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels_in_lavaan_parTable.R
\name{add_labels_in_lavaan_parTable}
\alias{add_labels_in_lavaan_parTable}
\title{Add Labels to \code{lavaan} \code{parTable}}
\usage{
add_labels_in_lavaan_parTable(internal_list = NULL)
}
\arguments{
\item{internal_list}{A list with various information extracted 
from the model.}
}
\value{
A \code{lavaan} \code{parTable} object where empty slots (i.e., 
parameters that do not have a user specified label) in the original 
\code{parTable} object have been filled in with default labels from 
the \code{lavaan} model syntax.
}
\description{
Fills in default labels from the \code{lavaan} model 
syntax into empty slots (i.e., parameters that do not have a user
specified label) of the lavaan parTable object.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
