% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ase_interventional_means.R
\name{calculate_ase_interventional_means}
\alias{calculate_ase_interventional_means}
\title{Calculate Asymptotics of Interventional Mean}
\usage{
calculate_ase_interventional_means(
  model = NULL,
  x = NULL,
  intervention_names = NULL,
  outcome_names = NULL,
  verbose = NULL
)
}
\arguments{
\item{model}{Object of class \code{causalSEM}.}

\item{x}{Numveric vector of interventional levels.}

\item{intervention_names}{Character vector of names of interventional 
variables.}

\item{outcome_names}{Character vector of names of outcome variables.}

\item{verbose}{Integer number setting verbosity of console outputs.}
}
\value{
The asymptotic covariance matrix, the asymptotic 
standard errors, and the approximate z-values of the the interventional
means for a specific interventional level. See, for example, Theorem 9 and 
Corollaries 10 and 11 in Gische and Voelkle (2021).
}
\description{
Calculates the asysmptotic covariance matrix,
the aysmptotic standard errors, and the approximate z-values 
of the of the interventional means for a specific interventional level.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
