% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interventional_density.R
\name{calculate_interventional_density}
\alias{calculate_interventional_density}
\title{Calculate PDF of Interventional Distribution}
\usage{
calculate_interventional_density(
  E = NULL,
  V = NULL,
  var_names = NULL,
  n_grid = NULL,
  model = NULL,
  use_model_values = FALSE,
  verbose = NULL
)
}
\arguments{
\item{E}{Numeric vector of mean values.}

\item{V}{Numeric vector of variances.}

\item{var_names}{Character vector of variable names.}

\item{verbose}{Integer number describing the verbosity of console output.
Admissible values: 0: no output (default), 1: user messages, 
2: debugging-relevant messages.}
}
\value{
List of numeric vectors (named if \code{var_names} are supplied) 
containing the values of the marginal pdf of each outcome variable (over 
a grid of values).
}
\description{
Calculate the probability density function (pdf) of the 
interventional distribution. The values are computed pointwise over a grid of
values for each univariate variable separately (i.e., the pdfs of the 
 marginal distributions). See, for example, Eqs. 9, 14, and 22c in Gische and
  Voelkle (2022).
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
