% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interventional_means.R
\name{calculate_interventional_means}
\alias{calculate_interventional_means}
\title{Calculate Interventional Mean}
\usage{
calculate_interventional_means(
  C = NULL,
  x = NULL,
  SI = NULL,
  n = NULL,
  IN = NULL,
  verbose = NULL
)
}
\arguments{
\item{C}{Numeric matrix of structural coefficients.}

\item{x}{Numeric vector of interventional levels.}

\item{SI}{Numeric selection matrix.}

\item{n}{Integer number of observed variables.}

\item{IN}{Numeric zero-one matrix.}

\item{verbose}{Integer number describing the verbosity of console output.
Admissible values: 0: no output (default), 1: user messages, 
2: debugging-relevant messages.}

\item{Psi}{Numeric covariance matrix (Psi-matrix).}
}
\value{
The mean vector of the interventional distribution.
}
\description{
Calculate mean vector of the interventional distribution 
(see, for example, Eq. 6a in Gische and Voelkle, 2022). The arguments of 
the \code{\link{calculate_interventional_means}} function are described in 
Definition 1 in Gische and Voelkle (2022).
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
