% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interventional_probabilities.R
\name{calculate_interventional_probabilities}
\alias{calculate_interventional_probabilities}
\title{Calculate Probabilities of Interventional Events}
\usage{
calculate_interventional_probabilities(
  mean = NULL,
  sd = NULL,
  y_low = NULL,
  y_up = NULL,
  model = NULL,
  use_model_values = FALSE,
  verbose = NULL
)
}
\arguments{
\item{mean}{Numeric vector of means.}

\item{sd}{Numeric vector of standard deviations.}

\item{y_low}{Numeric vector of lower bounds.}

\item{y_up}{Numeric vector of upper bounds.}

\item{verbose}{Integer number setting verbosity of console outputs.}
}
\value{
List of numeric vectors of probabilities interventional events.
}
\description{
Calculates probabilities of interventional events, for example,
the probability that the outcome variable realizes within a critical range of
values given an intervention on the exposure. See, for example, Eqs. 10, 15, 
and 22d in Gische and Voelkle (2022).
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
