% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jacobian_C.R
\name{calculate_jacobian_C}
\alias{calculate_jacobian_C}
\title{Calculate Jacobian of Matrix of Structural Coefficients}
\usage{
calculate_jacobian_C(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted internal_list with the slot "derivative" of
the ..$C slot containing the vectorized Jacobian of the C matrix.
}
\description{
Calculates the Jacobian of the vectorized matrix of structural 
coefficients (partial derivative with respect to the parameters vector theta)
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\seealso{
\code{\link{calculate_jacobian_Psi}}
}
