% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jacobian_Psi.R
\name{calculate_jacobian_Psi}
\alias{calculate_jacobian_Psi}
\title{Calculate Jacobian of Covariance Matrix}
\usage{
calculate_jacobian_Psi(internal_list = NULL)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted list with the slot \code{..$derivative}
 in the sublist \code{..$Psi} filled in:
\tabular{lll}{\code{..$derivative}:\tab \tab Numeric matrix containing 
     the Jacobian of the vectorized Psi matrix.}
}
\description{
Calculates the Jacobian of the vectorized covariance matrix of  
error terms (Psi-matrix). The Jacobian is defined as the partial derivative 
with respect to the parameters vector theta of distinct and functionally 
unrelated parameters.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
