% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.causalSEM.R
\name{coef.causalSEM}
\alias{coef.causalSEM}
\title{Extract Causal Coefficients}
\usage{
coef.causalSEM(x, what = "means")
}
\arguments{
\item{x}{an object of class "causalSEM", usually, a result of a call to intervention_effect.}

\item{what}{string indicating what type of coefficients are to be returned.
'means' returns interventional means, 'var' returns interventional variances,
and 'varcov' returns interventional variances and covariances as numeric
vectors. 'varcov_matrix' returns interventional variances and covariances
as a matrix.}
}
\value{
\code{coef.causalSEM} numeric vector or matrix.
}
\description{
coef method for class "causalSEM".
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
