% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_C.R
\name{fill_in_C}
\alias{fill_in_C}
\title{Fill Matrix of Structural Coefficients in List}
\usage{
fill_in_C(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted list with slots in the sublist \code{..$C} filled in:
\tabular{lll}{
     \code{..$values}:\tab \tab Numeric matrix containing 
     parameter values of matrix of structural coefficients.\cr
     \code{..$labels}:\tab \tab Character matrix containing 
     parameter labels of matrix of structural coefficients.\cr
     \tab \tab (NA for unlabeled parameters, NULL if labels can not 
     be extracted.)}
}
\description{
Extracts the structural coefficient matrix from a fitted
   structural equation model and fills them into the internal list.
   Supported objects types: \code{causalSEM}.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
