% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_C.R
\name{fill_in_C}
\alias{fill_in_C}
\title{Extracts the Structural Coefficient Matrix from a Fitted SEM Object}
\usage{
fill_in_C(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted internal_list with two slot in the sublist 
   C filled in: slot "values" is a numeric matrix containing the structural 
   coefficients (if available with row- and column names); slot "labels" is 
   a character matrix containing parameter labels (NA for unlabeled 
   parameters), NULL if labels are not extractable.
}
\description{
Extracts the structural coefficient matrix from a fitted
   structural equation model. Supported fitted objects: lavaan.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\seealso{
\code{\link{build_Psi}}
}
