% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_constant_matrices.R
\name{fill_in_constant_matrices}
\alias{fill_in_constant_matrices}
\title{Fill in Constant Matrices Used for Compuations to the Internal List}
\usage{
fill_in_constant_matrices(internal_list = NULL)
}
\arguments{
\item{internal_list}{A list with information extracted from the model.}
}
\value{

}
\description{
Fill in constant matrices used for the computation of the 
interventional distribution into the internal list.
(see for example Definition 1 in Gische and Voelkle, 2021)
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
