% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_info_interventions.R
\name{fill_in_info_interventions}
\alias{fill_in_info_interventions}
\title{Processes Arguments of the \code{intervention_effect}() Function.}
\usage{
fill_in_info_interventions(
  internal_list = NULL,
  intervention = NULL,
  outcome = NULL,
  intervention_level = NULL,
  effect_type = NULL,
  lower_bound = NULL,
  upper_bound = NULL
)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
Returns the inputted list with slots
   \code{n_intervention}, \code{intervention_names},
   \code{intervention_levels},
   \code{effect_type}, \code{n_outcome}, \code{outcome_names},
   \code{lower_bounds}, and \code{upper_bounds} filled in.
}
\description{
Extract interventional information from the arguments
of the \code{intervention_effect}() function.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
