% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_info_interventions.R
\name{fill_in_info_interventions}
\alias{fill_in_info_interventions}
\title{Extracts Interventional Information From the Arguments of the
intervention_effect() function.}
\usage{
fill_in_info_interventions(
  internal_list,
  intervention,
  outcome,
  intervention_level,
  effect_type,
  lower_bound,
  upper_bound
)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
Returns the inputted internal_list with slots
   \code{n_intervention}, \code{intervention_names},
   \code{intervention_levels},
   \code{effect_type}, \code{n_outcome}, \code{outcome_names},
   \code{lower_bounds}, and \code{upper_bounds} filled in.
}
\description{
Extract interventional information from the arguments
   of the intervention_effect() function.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible framework for
studying causal effects using linear models. Psychometrika
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\seealso{
\code{\link{populate_model_info}}
}
