% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_interventional_covariance_matrix.R
\name{fill_in_interventional_covariance_matrix}
\alias{fill_in_interventional_covariance_matrix}
\title{Fill in the Interventional Covariance Matrix to Internal List}
\usage{
fill_in_interventional_covariance_matrix(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted internal_list with slot
   interventional_distribution$moments$variance_matrix filled in.
}
\description{
Fills in slot internal_list$interventional_distribution$
covariance_matrix$values
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
