% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_interventional_density.R
\name{fill_in_interventional_density}
\alias{fill_in_interventional_density}
\title{Fill in Density Function of the Interventional Distribution}
\usage{
fill_in_interventional_density(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted internal_list with slot
internal_list$interventional_distribution$density_function$values filled in.
}
\description{
Fills in slot internal_list$interventional_distribution$
density_function$values in the internal_list.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\keyword{internal}
