% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_theta.R
\name{fill_in_theta}
\alias{fill_in_theta}
\title{Extracts Distinct and Functionally Unrelated SEM model Parameters}
\usage{
fill_in_theta(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted internal_list with slot ..$param filled in:
 "n_par" is an integer indicating the total number of estimated parameters
 (potential duplicates due to symmetry or equality constraints are counted).
 "n_par_unique" is an integer indicating the number of distinct and 
 functionally unrelated parameters.
 "labels_par_unique" is a character vector containing the labels of distinct
 and functionally unrelated parameters.
 "values_par_unique" is a numeric vector containing the parameter values
 (estimates) of distinct and functionally unrelated parameters.
 "varcov_par_unique" is a the (numeric) covariance
 matrix of the estimator of distinct and functionally unrelated parameters.
}
\description{
Rxtracts the numeric values and
  labels of distinct and functionally unrelated parameters from a
  fitted structural equation model (i.e., all uniquely labelled
  coefficients in C and Psi matrix. Duplicates due to symmetry or
  equality constraints are removed.). The entries in the resulting
  vectors follow the following predefined order:
  First: structural coefficients from the C matrix are added rowwise.
  Second: variance covariance parameters from the Psi matrix are added 
  rowwise.
  Supported fitted objects: lavaan.
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
\seealso{
\code{\link{fill_in_Psi, fill_in_C}}
}
\keyword{internal}
