% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_verbose_argument.R
\name{handle_verbose_argument}
\alias{handle_verbose_argument}
\title{Handles the Verbose Argument}
\usage{
handle_verbose_argument(verbose = NULL)
}
\arguments{
\item{verbose}{A single number, integer. 
0...no output (default); 
1...user messages;
2...debugging-relevant messages.}
}
\value{
Returns verbose setting.
}
\description{
Reads the user-inputted verbose
   argument checks if it is valid.
}
