% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_effect.R
\name{intervention_effect}
\alias{intervention_effect}
\title{Calculate Estimates of Causal Effects}
\usage{
intervention_effect(
  model = NULL,
  intervention = NULL,
  intervention_level = NULL,
  outcome = NULL,
  effect_type = NULL,
  lower_bound = NULL,
  upper_bound = NULL,
  verbose = 0,
  ...
)
}
\arguments{
\item{model}{Fitted model. The fitted model can be of class lavaan.}

\item{intervention}{Character vector. Variable names of interventional 
variables. Default: empty vector (no intervention).}

\item{intervention_level}{Numeric vector. Interventional levels. 
Same length and order as argument intervention. Default: vector of ones.}

\item{outcome}{Character vector. Variable names of outcome variables.
Default: all non-interventional variables.}

\item{effect_type}{Character vector. Features of the interventional 
distribution to be analyzed. Admissible values are 
"mean", "variance", "density", and "probability". 
Default: "mean".}

\item{lower_bound}{Numeric vector. Same length and order as argument outcome.
Lower bounds of critical range of outcome variables.}

\item{upper_bound}{Numeric vector. Same length and order as argument outcome.
Lower bounds of critical range of outcome variables.}

\item{verbose}{Single number, integer. 0...no output (default), 
1...user messages, 2...debugging-relevant messages.}
}
\value{
An object of class causalSEM, for which several methods 
are available, including a summary method.
}
\description{
Calculate estimates of causal effects based on the 
interventional distribution.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
\keyword{external}
