% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_effect.R
\name{intervention_effect}
\alias{intervention_effect}
\title{Calculate Estimates of Causal Effects}
\usage{
intervention_effect(
  model = NULL,
  intervention = NULL,
  intervention_level = NULL,
  outcome = NULL,
  effect_type = NULL,
  lower_bound = NULL,
  upper_bound = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{model}{Fitted model. The fitted model can be of class \code{lavaan}.}

\item{intervention}{Character vector of names of interventional 
variables.}

\item{intervention_level}{Numeric vector of interventional levels. 
Same length and order as argument \code{intervention}. 
Default: vector of ones.}

\item{outcome}{Character vector of variable names of outcome variables.
Default: all non-interventional variables.}

\item{effect_type}{Character string describing the features of the 
interventional distribution to be analyzed. Admissible values are 
\code{mean} (default), \code{variance}, \code{density}, and 
\code{probability}.}

\item{lower_bound}{Numeric vector of same length and order as argument 
\code{outcome}. Lower bounds of critical range of outcome variables.}

\item{upper_bound}{Numeric vector of same length and order as argument
\code{outcome}. Upper bounds of critical range of outcome variables.}

\item{verbose}{Integer number describing the verbosity of console output.
Admissible values: 0: no output (default), 1: user messages, 
2: debugging-relevant messages.}
}
\value{
An object of class \code{causalSEM} for which several methods 
are available including \code{\link{summary.causalSEM}} and 
\code{\link{print.causalSEM}}.
}
\description{
Calculate estimates of causal effects based on the 
interventional distribution.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
\keyword{external}
