% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_empty_list.R
\name{make_empty_list}
\alias{make_empty_list}
\title{Builds an empty internal list}
\usage{
make_empty_list(verbose = NULL)
}
\arguments{
\item{verbose}{A number, 0...no output (default), 1...user messages,
2...debugging-relevant messages. Valid default is
set by function \code{\link{verbose_argument_handling}}}
}
\value{
\code{make_empty_list} returns an empty internal list with structure:\cr
 \code{\tabular{lll}{
    List of 5\cr
    $ fitted_object      : NULL           \tab \tab \cr
    $ fitted_object_class: NULL           \tab \tab \cr
    $ info_model         :List of 10      \tab \tab \cr
     ..$ n_obs            : int 0         \tab \tab number of observations\cr
     ..$ n_ov             : int 0         \tab \tab number of manifest variables\cr
     ..$ var_names        : chr(0)        \tab \tab names of observed variables\cr
     ..$ C                :List of 2      \tab \tab \cr
     .. ..$ values: num[0, 0]           \tab \tab parameter values of matrix of structural coefficients\cr
     .. ..$ labels: chr[0, 0]           \tab \tab parameter names of matrix of structural coefficients\cr
     .. ..$ derivative: chr[0, 0]       \tab \tab partial derivative of the vectorized C matrix with respect to the parameters\cr
     ..$ Psi              :List of 2      \tab \tab \cr
     .. ..$ values: num[0, 0]           \tab \tab parameter values of model implied covariance matrix\cr
     .. ..$ labels: chr[0, 0]           \tab \tab parameter names of model implied covariance matrix\cr
     .. ..$ derivative: chr[0, 0]       \tab \tab partial derivative of the vectorized Psi matrix with respect to the parameters\cr
     ..$ param            :List of 5      \tab \tab \cr
     .. ..$ n_par            : int 0         \tab \tab total number of estimated parameters\cr
     .. ..$ n_par_unique     : int 0         \tab \tab number of distinct and functionally unrelated parameters\cr
     .. ..$ labels_par_unique : chr(0)        \tab \tab names of distinct and functionally unrelated parameters\cr
     .. ..$ values_par_unique : num(0)        \tab \tab parameter values (estimates) of distinct and functionally unrelated parameters\cr
     .. ..$ varcov_par_unique: num[0, 0]   \tab \tab variance-covariance matrix of the estimator of distinct and functionally unrelated parameters\cr
    $ info_interventions : List of 8       \tab \tab \cr
     ..$ n_intervention    : int 0        \tab \tab number of interventional variables\cr
     ..$ intervention_name : chr(0)       \tab \tab names of interventional variables\cr
     ..$ n_outcome         : int 0        \tab \tab number of outcome variables\cr
     ..$ outcome_name      : chr(0)       \tab \tab names of outcome variables\cr
     ..$ intervention_level: num(0)       \tab \tab interventional level\cr
     ..$ effect_type       : chr(0)       \tab \tab parts of the interventional distribution the user is interested in\cr
     ..$ lower_bound       : num(0)       \tab \tab lower bound of critical range of univariate outcome variable\cr
     ..$ upper_bound       : num(0)       \tab \tab upper bound of critical range of univariate outcome variable\cr
    $ interventional_distribution : List of 4 \tab \tab \cr
     ..$ zero_one_matrices : List of 3    \tab \tab \cr
     .. .. $ select_intervention: num[0, 0]  \tab \tab selection matrix for entries that are intervened on \cr
     .. .. $ select_non_intervention: num[0, 0]  \tab \tab selection matrix for entries that are NOT intervened on \cr
     .. .. $ eliminate_intervention: num[0, 0]  \tab \tab matrix that replaces entries that are intervened on by zero \cr
     ..$ moments : List of 2    \tab \tab \cr
     .. .. $ mean_vector: num(0)  \tab \tab mean vector of the interventional distribution  \cr
     .. .. $ variance_matrix: num[0, 0]  \tab \tab variance-covariance matrix of the interventional distribution \cr
     ..$ density : List of 1    \tab \tab \cr
     .. .. $ pdf: num[0, 0]  \tab \tab probability density function interventional distribution \cr
     ..$ probability : List of 1    \tab \tab \cr
     .. .. $ p: num[0, 0]  \tab \tab probability of interventional event \cr
    $ control            : List of 1      \tab \tab \cr
     ..$ verbose: num(0)                   \tab \tab verbosity of console output\cr
 }}
}
\description{
Internal function that builds an empty internal list;
   verbose argument is written into the internal list
}
\references{
Gische, C. & Voelkle, M. C. (under review). Beyond the mean: A flexible framework for
   studying causal effects using linear models. \url{https://www.researchgate.net/profile/Christian-Gische/publication/335030449_Gische_Voelkle_Causal_Inference_in_Linear_Models/links/6054eb6e299bf1736755110b/Gische-Voelkle-Causal-Inference-in-Linear-Models.pdf}
}
\seealso{
\code{\link{verbose_argument_handling}}
}
\keyword{internal}
