% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_table_means.R
\name{prepare_table_means}
\alias{prepare_table_means}
\title{Prepare Tables with Information About Interventional Means}
\usage{
prepare_table_means(x, digits = 3)
}
\arguments{
\item{x}{an object of class "causalSEM", usually, a result of a call to
intervention_effect.}

\item{digit}{integer indicating the number of decimal places (round) or
significant digits (signif) to be used.}
}
\value{
\code{prepare_table_means} returns a character vector.
}
\description{
Internal helper function that creates a formatted table
containing various information about the interventional means. Used in
print.causalSEM and summary.causalSEM
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible
framework for studying causal effects using linear models. Psychometrika
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
