% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.causalSEM.R
\name{print.causalSEM}
\alias{print.causalSEM}
\title{Printing a Causal SEM Object}
\usage{
\method{print}{causalSEM}(x, digits = 3)
}
\arguments{
\item{x}{an object of class "causalSEM", usually, a result of a call to intervention_effect.}

\item{digit}{integer indicating the number of decimal places (round) or
significant digits (signif) to be used.}
}
\value{
\code{print.causalSEM} returns a table with summary statistics.
}
\description{
print method for class "causalSEM".
}
\references{
Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible 
framework for studying causal effects using linear models. Psychometrika 
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
