% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.causalSEM.R
\name{print.causalSEM}
\alias{print.causalSEM}
\title{Print Object of Class \code{causalSEM}}
\usage{
\method{print}{causalSEM}(x, digits = 3)
}
\arguments{
\item{x}{An object of class \code{causalSEM}.}

\item{digit}{Single number, integer. Indicating the number of decimal places
(round) or significant digits (signif) to be used.}
}
\value{
A table summarizing selected information and statistics from an 
object of class \code{causalSEM}.
}
\description{
Print method for objects of class \code{causalSEM}.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
