% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.causalSEM.R
\name{summary.causalSEM}
\alias{summary.causalSEM}
\title{Summarizing a Causal SEM Object}
\usage{
\method{summary}{causalSEM}(x, digits = 3)
}
\arguments{
\item{x}{an object of class "causalSEM", usually, a result of a call to intervention_effect.}

\item{digit}{integer indicating the number of decimal places (round) or
significant digits (signif) to be used.}
}
\value{
\code{summary.causalSEM} prints a list of summary statistics.
}
\description{
summary method for class "causalSEM".
}
\references{


Gische, C., Voelkle, M.C. (2021) Beyond the mean: a flexible
framework for studying causal effects using linear models. Psychometrika
(advanced online publication). https://doi.org/10.1007/s11336-021-09811-z
}
